seperated logic for external file updation
