+++ scripts/storage-api.js
